+++ CHANGELOG.md
@@ -13,6 +13,7 @@ All notable changes to this project will be documented in this file.
 * Fix comment input area not being visible in edit mode sometimes
 * Fix engine color indicator of GTP console being wrong when swapping engine color mid-game
 * Up and down keys for traversing command history in the GTP console are working again
+* Fix Sabaki trying to add engine variations while in scoring mode after a double pass
 
 ## [Sabaki v0.32.2][v0.32.2] (2018-01-01)